refactor: updating `markViral` to use less gas & updated `Pool` info

+++ contracts/interfaces/IStoreable.sol
@@ -89,7 +89,7 @@ interface IStoreable {
         uint256 amount;
         uint256 shares;
         uint256 created;
-        uint8 viralityScore;
+        uint64 viralityScore;
     }
 
     /**
